UXD-331 Fix: adjustments for variations and docs
